+++ core/api_container/api_container_env_var_values/api_container_params_json/test_execution_args.go
@@ -17,6 +17,32 @@ const (
 	runTimeoutFieldname   = "testRunTimeout"
 )
 
+// Params that the API container will use to construct the free host port binding supplier that it will use
+// for binding service ports to host ports for debugging purposes
+type HostPortBindingSupplierParams struct {
+	// The host interface IP that the doled-out ports should be on
+	InterfaceIp		string `json:"interfaceIp"`
+
+	// The protocol of the ports that the supplier will produce
+	Protocol		string `json:"protocol"`
+
+	// Inclusive start of port range that will be doled out
+	PortRangeStart	uint32 `json:"portRangeStart"`
+
+	// EXCLUSIVE end of port range that will be doled out
+	PortRangeEnd 	uint32	`json:"portRangeEnd"`
+
+	// "Set" of ports that are already taken and shouldn't be doled out by the free host port binding supplier
+	// NOTE: This is more a hint, rather than a requirement, since the FreeHostPortBindingSupplier that's constructed
+	//  from these params will test if a port is free regardless
+	TakenPorts map[uint32]bool	`json:"takenPorts"`
+}
+
+// Even though the fields are public, we create & use this constructor so that new fields will cause an API break
+func NewHostPortBindingSupplierParams(interfaceIp string, protocol string, portRangeStart uint32, portRangeEnd uint32, takenPorts map[uint32]bool) *HostPortBindingSupplierParams {
+	return &HostPortBindingSupplierParams{InterfaceIp: interfaceIp, Protocol: protocol, PortRangeStart: portRangeStart, PortRangeEnd: portRangeEnd, TakenPorts: takenPorts}
+}
+
 // Fields are public for JSON de/serialization
 type TestExecutionArgs struct {
 	ExecutionInstanceId      string	`json:"executionInstanceId"`
@@ -38,6 +64,11 @@ type TestExecutionArgs struct {
 	// TODO remove this by passing over the test metadata as one of the params, so that the API container
 	//  knows about the metadata
 	IsPartitioningEnabled bool	`json:"isPartitioningEnabled"`
+
+	// A non-nil value indicates that the Kurtosis API container should bind service ports to ports on the
+	// host machine running Kurtosis, for debugging purposes
+	// If this is nil, no host port binding will occur
+	HostPortBindingSupplierParams *HostPortBindingSupplierParams `json:"hostPortBindingSupplierParams"`
 }
 
 // Even though the fields are public due to JSON de/serialization requirements, we still have this constructor so that
@@ -54,7 +85,8 @@ func NewTestExecutionArgs(
 		apiContainerIpAddr string,
 		testSetupTimeoutInSeconds uint32,
 		testRunTimeoutInSeconds uint32,
-		isPartitioningEnabled bool) (*TestExecutionArgs, error) {
+		isPartitioningEnabled bool,
+		hostPortBindingSupplierParams *HostPortBindingSupplierParams) (*TestExecutionArgs, error) {
 	result := TestExecutionArgs{
 		ExecutionInstanceId:       executionInstanceId,
 		NetworkId:                 networkId,
@@ -68,6 +100,7 @@ func NewTestExecutionArgs(
 		TestSetupTimeoutInSeconds: testSetupTimeoutInSeconds,
 		TestRunTimeoutInSeconds:   testRunTimeoutInSeconds,
 		IsPartitioningEnabled:     isPartitioningEnabled,
+		HostPortBindingSupplierParams: hostPortBindingSupplierParams,
 	}
 	if err := result.validate(); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred validating test execution args")